Fix transfer-encoding vs. content-length, and provide passthru mode.
transfer-encoding = chunked vs. content-length encoding is documented
in rfc 2616 3.6.1 and rfc 7230 4.1.  In both cases, it is
documented that if transfer-encoding == 'chunked', then content-length
should not be supplied by a client, and must not be processed by a server.

The prevoius code in civetweb allowed content-length to override
transfer-encoding=chunked, which is clearly wrong by the standard.

Additionally, for ceph, when doing awsv4 "streaming" signatures,
there is a per-chunk signature that is recorded as a special field
inside of each chunk header.  There are actually two modes for s3
chunked transfers, one using "transfer-encoding: chunked", which is
not well documented, and another using "content-encoding: aws-chunked",
which is the preferred method.  Since in the later case ceph has
to have de-chunking logic anyways, it is less code to maintain
and debug to have ceph always do the chunk decoding, hence,
this commit contains an optional "pass-through" mode to make this possible.

Signed-off-by: Marcus Watts <mwatts@redhat.com>