Fix transfer-encoding: chunked encoding to work with more clients.
In rfc 2616 3.6.1 and rfc 7230 4.1, a mechanism is provided to supply
extra data in chunk headers, in the of keyword/value pairs.  Old versions
of civetweb completely ignored this data.

The boto2 s3 client, when it uses chunked encoding, sends a per-chunk
header consisting of a hex count followed by a single ';'.  This is not,
according to a strict reading of the rfcs, legal, but it is something that
ought to be acceptable to any server that does not expect to interpret
any keyword values.  This behavior is not likely to change.

Additionally, the final chunk header is supposed to be followed by not
one, but two crlf pairs.  It is desirable to consume both pairs rather
than just leaving the 2nd pair to cause trouble later.

Older versions of civetweb did no checking of any possible chunk header
data, they only checked to see if they could decode a hexadecimal count.
This is clearly more generous than necessary, but the change to be more
strict broke boto2 clinets, which is not desirable.

This change provides slightly more general logic that allows for a ';'
followed by optional data (which is unconditionally discarded for now),
and also consumes the extra crlf pair.

Signed-off-by: Marcus Watts <mwatts@redhat.com>