Fix transfer-encoding: chunked encoding to work with more clients.
The final chunk header of size 0 is followed by two crlf pairs.
The first ends the chunk header, and is properly digested by
civetweb.  The 2nd ends the "dummy" datablock after the last
chunk header.  It must be consumed as well, or else the
next request on this connection will fail.  This change adds
logic to properly digest the 2nd crlf pair.

Signed-off-by: Marcus Watts <mwatts@redhat.com>