Fix transfer-encoding vs. content-length.
transfer-encoding = chunked vs. content-length encoding is documented
in rfc 2616 3.6.1 and rfc 7230 4.1.  In both cases, it is
documented that if transfer-encoding == 'chunked', then content-length
should not be supplied by a client, and must not be processed by a server.

The previous code in civetweb allowed content-length to override
transfer-encoding=chunked, which is clearly wrong by the standard.
This chnage corrects the order of those checks.

Signed-off-by: Marcus Watts <mwatts@redhat.com>