Provide new interface mg_set_must_close
In some cases, ceph may detect an error that indicates
that further use of this connection is doubtful.  For instance,
mg_read() fails to return expected data, due to a chunk decoding
error.  In this case, the connection should be closed, but not
until after the error is sent.  civetweb provides low-level logic
to do this (conn->must_close); this commit provides an API
for the application to trigger that behavior.

Signed-off-by: Marcus Watts <mwatts@redhat.com>